Move configuration code out of tree.ts
